+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Note that since we don't clearly distinguish between a public and private interf
     - Remove dMarkerType shader define (use uMarker as needed)
     - Support to ignore defines depending on the shader variant
     - Combine pickObject/pickInstance/pickGroup shader variants into one
+    - Correctly set shader define flags for overpaint, transparency, substance, clipping
 
 ## [v3.0.0-dev.5] - 2021-12-16